refactor: ref 관련 유틸 함수 /utils 폴더로 분리
